+++ core-web/apps/dotcms-ui/src/app/portlets/dot-apps/components/dot-apps-configuration-detail/components/dot-apps-configuration-detail-generated-string-field/dot-apps-configuration-detail-generated-string-field.component.spec.ts
