+++ core-web/apps/dotcms-ui/src/app/portlets/dot-apps/components/dot-apps-configuration-detail/components/dot-apps-configuration-header/dot-apps-configuration-header.component.spec.ts
@@ -14,7 +14,7 @@ import { MockDotMessageService, MockDotRouterService } from '@dotcms/utils-testi
 
 import { DotAppsConfigurationHeaderComponent } from './dot-apps-configuration-header.component';
 
-import { DotCopyLinkComponent } from '../../../view/components/dot-copy-link/dot-copy-link.component';
+import { DotCopyLinkComponent } from '../../../../../../view/components/dot-copy-link/dot-copy-link.component';
 
 @Component({
     // eslint-disable-next-line @angular-eslint/component-selector
