+++ core-web/apps/dotcms-ui/src/app/portlets/dot-apps/components/dot-apps-configuration-detail/dot-apps-configuration-detail.component.html
@@ -1,8 +1,6 @@
 <div class="dot-apps-configuration-detail__container">
     <div class="dot-apps-configuration-detail__header">
-        @if (apps) {
-            <dot-apps-configuration-header [app]="apps" />
-        }
+        <dot-apps-configuration-header [app]="apps" />
         <div class="dot-apps-configuration-detail__host-name">
             <span>{{ apps.sites[0].name }}</span>
             <div class="dot-apps-configuration-detail-actions">
